+++ lib/draftking/queue.rb
@@ -8,9 +8,10 @@ module DK
       options[:message] = 'Moving Queue ~> Drafts: '
       options[:shuffle] = false
       options[:state]   = DK::DRAFT
-      post_operation(options) do |post, _|
+      mod_count, mod_posts = post_operation(options) do |post, _|
         post.changed = !post.has_key_text?(key_text: @key_text)
       end
+      mod_count
     end
   end
 end